move connection methods (for button slots and etc) to special file
